refactor eplb logic to base MoE class
